requirements.txt
aniso8601==9.0.1
attrs==21.4.0
cachetools==4.2.4
certifi==2022.12.7
charset-normalizer==2.0.12
click==8.1.2
clickclick==20.10.2
colorama==0.4.4
dlib @ file:///D:/bld/dlib_1652035369210/work
efficientnet-pytorch==0.7.1
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.9.0
Flask==2.1.1
Flask-RESTful==0.3.9
flask-restplus==0.13.0
google-api-core==2.10.2
google-auth==1.35.0
google-cloud-core==1.7.3
google-cloud-storage==1.16.1
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.58.0
huggingface-hub==0.11.1
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.0.3
joblib==1.2.0
jsonschema==4.4.0
MarkupSafe==2.1.1
munch==2.5.0
numpy @ file:///D:/bld/numpy_1651020737746/work
opencv-python==4.7.0.68
packaging==21.3
Pillow==9.1.0
pretrainedmodels==0.7.4
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.8
pyrsistent==0.18.1
python-utils==3.4.5
pytz==2022.1
PyYAML==6.0
requests==2.27.1
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.1
segmentation-models-pytorch==0.3.2
six==1.16.0
sklearn==0.0.post1
threadpoolctl==3.1.0
timm==0.6.12
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.1
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.0.0
wincertstore==0.2
zipp==3.8.0
